requirements.txt
# Analytics Service Dependencies

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Technical analysis
pandas-ta>=0.3.14b0

# Kafka
kafka-python>=2.0.0

# Database
psycopg2-binary>=2.9.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
